feat: Add weekdays to onboarding
- Add training days to onboarding request struct as a string so that
frontend doesn't send magic numbers.
- Create WeekdayTranslation to convert string values to ints.
- Add TrainingDays to User struct to save them to database.
- Utilize "comma ok" idiom for to check if event provides points or not.
- Add some important TODOs.
- Replace int64 with int.

+++ src/internal/progress/domain/repository.go
@@ -53,7 +53,7 @@ type EventLogRepository interface {
 // LeaderboardRepository defines the contract for leaderboard operations.
 type LeaderboardRepository interface {
 	// UpdateScore updates a user's score on the leaderboard.
-	UpdateScore(ctx context.Context, userID string, points int64) error
+	UpdateScore(ctx context.Context, userID string, points int) error
 
 	// GetTopPlayers retrieves the top players for the leaderboard.
 	GetTopPlayers(ctx context.Context, offset, limit int) ([]LeaderboardEntry, error)
